Added more branch options to enforce-branch-name-rules.yml (#4636)
I think it makes sense to support also
fe - for just front end changes (recently had something like that for an RDI fix) in which cases there is no point in running the BE and integrations tests
be - for just api changes. It also happens from time to time and it doesn't make sense to run all of our FE tests, especially how flaky they are.
e2e - just for e2e tests. No point in wasting a lot of time (physical and github) to run all of the other tests